requirements.txt
pandas>=1.0.0
python-dotenv>=0.19.0
openai>=0.27.0
numpy>=1.20.0
scikit-learn>=1.0.0
nltk>=3.6.0
matplotlib>=3.0.0
seaborn>=0.11.0
tqdm>=4.0.0
umap-learn>=0.5.0